requirements.txt
flask
flask-cors
sendgrid
python-dotenv
gunicorn
